fix hide drawer on call start, as well as refactor views to their own files
