+++ AzureCommunicationUI/sdk/AzureCommunicationUICalling/Sources/Presentation/SwiftUI/ViewComponents/Drawer/DrawerListView.swift
@@ -91,200 +91,6 @@ internal struct DrawerListSection {
     let items: [BaseDrawerItemViewModel]
 }
 
-internal struct SelectableDrawerItemView: View {
-    let item: SelectableDrawerListItemViewModel
-
-    var body: some View {
-        HStack {
-            Icon(name: item.icon, size: DrawerListConstants.iconSize)
-                .foregroundColor(.primary)
-
-            Text(item.title)
-                .foregroundColor(.primary)
-                .padding(.leading, DrawerListConstants.textPaddingLeading)
-                .font(.body)
-            Spacer()
-            if item.isSelected {
-                Icon(name: .checkmark, size: DrawerListConstants.iconSize)
-            }
-        }
-        .padding(.horizontal, DrawerListConstants.optionPaddingHorizontal)
-        .padding(.vertical, DrawerListConstants.optionPaddingVertical)
-        .frame(maxWidth: .infinity)
-        .contentShape(Rectangle())
-        .onTapGesture {
-            item.action()
-        }
-        .accessibilityIdentifier(item.accessibilityIdentifier)
-    }
-}
-
-internal struct DrawerItemView: View {
-    let item: DrawerListItemViewModel
-
-    var body: some View {
-        HStack {
-            if let icon = item.startIcon {
-                Icon(name: icon, size: DrawerListConstants.iconSize)
-                    .foregroundColor(.primary)
-            }
-            Text(item.title)
-                .foregroundColor(.primary)
-                .padding(.leading, DrawerListConstants.textPaddingLeading)
-                .font(.body)
-            Spacer()
-        }
-        .padding(.horizontal, DrawerListConstants.optionPaddingHorizontal)
-        .padding(.vertical, DrawerListConstants.optionPaddingVertical)
-        .frame(maxWidth: .infinity)
-        .contentShape(Rectangle())
-        .onTapGesture {
-            if let action = item.action {
-                action()
-            }
-        }
-        .accessibilityIdentifier(item.accessibilityIdentifier)
-    }
-}
-
-internal struct DrawerTitleView: View {
-    let item: TitleDrawerListItemViewModel
-
-    var body: some View {
-        HStack {
-            Spacer()
-            Text(item.title)
-                .foregroundColor(.primary)
-                .padding(.leading, DrawerListConstants.textPaddingLeading)
-                .font(.headline)
-            Spacer()
-        }
-        .padding(.horizontal, DrawerListConstants.optionPaddingHorizontal)
-        .padding(.vertical, DrawerListConstants.optionPaddingVertical)
-        .frame(maxWidth: .infinity)
-        .contentShape(Rectangle())
-        .accessibilityIdentifier(item.accessibilityIdentifier)
-        .background(Color(StyleProvider.color.surface))
-    }
-}
-
-internal struct DrawerBodyTextView: View {
-    let item: BodyTextDrawerListItemViewModel
-
-    var body: some View {
-        HStack {
-            Text(item.title)
-                .foregroundColor(.primary)
-                .padding(.leading, DrawerListConstants.textPaddingLeading)
-                .font(.body)
-            Spacer()
-        }
-        .padding(.horizontal, DrawerListConstants.optionPaddingHorizontal)
-        .padding(.vertical, DrawerListConstants.optionPaddingVertical)
-        .frame(maxWidth: .infinity)
-        .contentShape(Rectangle())
-        .accessibilityIdentifier(item.accessibilityIdentifier)
-        .background(Color(StyleProvider.color.surface))
-    }
-}
-
-internal struct DrawerBodyWithActionTextView: View {
-    let item: BodyTextWithActionDrawerListItemViewModel
-    @State
-    var isConfirming = false
-
-    var body: some View {
-        HStack {
-            Text(item.title)
-                .foregroundColor(.primary)
-                .padding(.leading, DrawerListConstants.textPaddingLeading)
-                .font(.body)
-            Spacer()
-            Text(item.actionText)
-                .onTapGesture {
-                   isConfirming = true
-                }
-                .foregroundColor(Color(StyleProvider.color.primaryColor))
-        }
-        .padding(.horizontal, DrawerListConstants.optionPaddingHorizontal)
-        .padding(.vertical, DrawerListConstants.optionPaddingVertical)
-        .frame(maxWidth: .infinity)
-        .contentShape(Rectangle())
-        .accessibilityIdentifier(item.accessibilityIdentifier)
-        .background(Color(StyleProvider.color.surface))
-        .confirmationDialog(
-            Text(item.confirmAccept),
-            isPresented: $isConfirming,
-            titleVisibility: .automatic) {
-            Button(item.actionText) {
-                isConfirming = false
-                item.accept()
-            }.foregroundColor(Color(StyleProvider.color.primaryColor))
-            Button(item.confirmDeny, role: .destructive) {
-                isConfirming = false
-                item.deny()
-            }
-            Button("Cancel", role: .cancel) {
-                isConfirming = false
-            }.foregroundColor(Color(StyleProvider.color.primaryColor))
-        }
-    }
-}
-
-struct BackgroundCleanerView: UIViewRepresentable {
-    func makeUIView(context: Context) -> UIView {
-        let view = UIView()
-        DispatchQueue.main.async {
-            view.superview?.superview?.backgroundColor = .clear
-        }
-        return view
-    }
-
-    func updateUIView(_ uiView: UIView, context: Context) {}
-}
-
-struct CustomAlert<Content: View>: View {
-    let title: String
-    let dismiss: () -> Void
-    let content: Content
-
-    init(title: String,
-         dismiss: @escaping () -> Void,
-         @ViewBuilder content: () -> Content) {
-        self.title = title
-        self.content = content()
-        self.dismiss = dismiss
-    }
-
-    var body: some View {
-            ZStack {
-                Color.black.opacity(0.5)
-                    .frame(maxWidth: .infinity, maxHeight: .infinity)
-                    .onTapGesture {
-                        dismiss()
-                    }
-                HStack {
-                    Spacer()
-                    VStack {
-                        Spacer()
-                        VStack {
-                            Text(title).font(.headline)
-                            HStack {
-                                content
-                            }
-                        }.padding()
-                        .background(Color(StyleProvider.color.surface))
-                        .cornerRadius(10)
-                        .shadow(radius: 10)
-
-                        Spacer()
-                    }
-                    Spacer()
-                }
-            }
-    }
-}
-
 internal class DrawerListConstants {
     static let iconSize: CGFloat = 24
     static let textPaddingLeading: CGFloat = 8
