+++ AzureCommunicationUI/sdk/AzureCommunicationUICalling/Sources/Presentation/SwiftUI/ViewComponents/Drawer/Views/DrawerBodyTextView.swift
@@ -1,0 +1,70 @@
+//
+//  Copyright (c) Microsoft Corporation. All rights reserved.
+//  Licensed under the MIT License.
+//
+
+import Foundation
+import SwiftUI
+
+internal struct DrawerBodyTextView: View {
+    let item: BodyTextDrawerListItemViewModel
+
+    var body: some View {
+        HStack {
+            Text(item.title)
+                .foregroundColor(.primary)
+                .padding(.leading, DrawerListConstants.textPaddingLeading)
+                .font(.body)
+            Spacer()
+        }
+        .padding(.horizontal, DrawerListConstants.optionPaddingHorizontal)
+        .padding(.vertical, DrawerListConstants.optionPaddingVertical)
+        .frame(maxWidth: .infinity)
+        .contentShape(Rectangle())
+        .accessibilityIdentifier(item.accessibilityIdentifier)
+        .background(Color(StyleProvider.color.surface))
+    }
+}
+
+internal struct DrawerBodyWithActionTextView: View {
+    let item: BodyTextWithActionDrawerListItemViewModel
+    @State
+    var isConfirming = false
+
+    var body: some View {
+        HStack {
+            Text(item.title)
+                .foregroundColor(.primary)
+                .padding(.leading, DrawerListConstants.textPaddingLeading)
+                .font(.body)
+            Spacer()
+            Text(item.actionText)
+                .onTapGesture {
+                    isConfirming = true
+                }
+                .foregroundColor(Color(StyleProvider.color.primaryColor))
+        }
+        .padding(.horizontal, DrawerListConstants.optionPaddingHorizontal)
+        .padding(.vertical, DrawerListConstants.optionPaddingVertical)
+        .frame(maxWidth: .infinity)
+        .contentShape(Rectangle())
+        .accessibilityIdentifier(item.accessibilityIdentifier)
+        .background(Color(StyleProvider.color.surface))
+        .confirmationDialog(
+            Text(item.confirmAccept),
+            isPresented: $isConfirming,
+            titleVisibility: .automatic) {
+                Button(item.actionText) {
+                    isConfirming = false
+                    item.accept()
+                }.foregroundColor(Color(StyleProvider.color.primaryColor))
+                Button(item.confirmDeny, role: .destructive) {
+                    isConfirming = false
+                    item.deny()
+                }
+                Button("Cancel", role: .cancel) {
+                    isConfirming = false
+                }.foregroundColor(Color(StyleProvider.color.primaryColor))
+        }
+    }
+}
