+++ AzureCommunicationUI/sdk/AzureCommunicationUICalling/Sources/Presentation/SwiftUI/ViewComponents/Drawer/Views/DrawerTitleView.swift
@@ -1,0 +1,27 @@
+//
+//  Copyright (c) Microsoft Corporation. All rights reserved.
+//  Licensed under the MIT License.
+//
+
+import Foundation
+import SwiftUI
+internal struct DrawerTitleView: View {
+    let item: TitleDrawerListItemViewModel
+
+    var body: some View {
+        HStack {
+            Spacer()
+            Text(item.title)
+                .foregroundColor(.primary)
+                .padding(.leading, DrawerListConstants.textPaddingLeading)
+                .font(.headline)
+            Spacer()
+        }
+        .padding(.horizontal, DrawerListConstants.optionPaddingHorizontal)
+        .padding(.vertical, DrawerListConstants.optionPaddingVertical)
+        .frame(maxWidth: .infinity)
+        .contentShape(Rectangle())
+        .accessibilityIdentifier(item.accessibilityIdentifier)
+        .background(Color(StyleProvider.color.surface))
+    }
+}
