+++ AzureCommunicationUI/sdk/AzureCommunicationUICalling/Sources/Presentation/SwiftUI/ViewComponents/Drawer/Views/SelectableDrawerItemView.swift
@@ -1,0 +1,35 @@
+//
+//  Copyright (c) Microsoft Corporation. All rights reserved.
+//  Licensed under the MIT License.
+//
+
+import Foundation
+import SwiftUI
+
+internal struct SelectableDrawerItemView: View {
+    let item: SelectableDrawerListItemViewModel
+
+    var body: some View {
+        HStack {
+            Icon(name: item.icon, size: DrawerListConstants.iconSize)
+                .foregroundColor(.primary)
+
+            Text(item.title)
+                .foregroundColor(.primary)
+                .padding(.leading, DrawerListConstants.textPaddingLeading)
+                .font(.body)
+            Spacer()
+            if item.isSelected {
+                Icon(name: .checkmark, size: DrawerListConstants.iconSize)
+            }
+        }
+        .padding(.horizontal, DrawerListConstants.optionPaddingHorizontal)
+        .padding(.vertical, DrawerListConstants.optionPaddingVertical)
+        .frame(maxWidth: .infinity)
+        .contentShape(Rectangle())
+        .onTapGesture {
+            item.action()
+        }
+        .accessibilityIdentifier(item.accessibilityIdentifier)
+    }
+}
